Support multiple trust anchors with same SubjectDN
Resolving the trust anchor for a certificate would fail randomly if having
several trust anchors with same Subject DN. It would pick the first
trust anchor with matching subject, and throw exception if the signature
validation fails. In the event of signature validation failing, it will
continue to search for matching trust anchors.