Replace Buypass' SEID2 root certs for test
Long story. We originally acquired the certs they published
valid from nov. 3rd 2020, but they got replaced shortly after.
https://github.com/digipost/api-client-shared-dotnet/pull/30#issuecomment-1875426635

This is possible due to the root certificate, which is self-signed, is created from the same keys,
and the already issued intermediate certs will still validate as issued
by the (new) root cert. This is OK, due to these certs are solely for
use for testing, and does not constitute any authority for actual
identification of parties.